refactored out logforth
