+++ src/lib.rs
@@ -9,6 +9,7 @@ pub mod all2all;
 pub mod consensus;
 pub mod crypto;
 pub mod disseminator;
+pub mod logging;
 pub mod network;
 pub mod repair;
 pub mod shredder;
